dfdf

+++ game.py
@@ -9,8 +9,8 @@ player = pygame.image.load("resources/images/dude.png")
 grass = pygame.image.load("resources/images/grass.png")
 castle = pygame.image.load("resources/images/castle.png")
 
-for x in range(width/grass.get_width()+1):
-    for y in range(height/grass.get_height()+1):
+for x in range((int)(width/grass.get_width()+1)):
+    for y in range((int)(height/grass.get_height()+1)):
         screen.blit(grass,(x*100,y*100))
 screen.blit(castle,(0,30))
 screen.blit(castle,(0,135))
